Use `CountingIdlingResource` to prevent UI test failures
To prevent UI tests from failing due to background tasks that have not made the data available yet, it was decided to follow best practices and use `CountingIdlingResource`. This means that, in production code, we need to track when a background test is running so that information can be used when running tests.