Refactor `Stat` class
Ideally we would like to maintain a sealed trait and data class hierarchy for `Stat` but it would be impossible (as far as I know) to have that with Room. For that reason, `Stat` was simplified to a single class.

This commit also introduces the decision of representing `Array` as a stringified JSON array, in the DB.