Create a view for `INT` stat type
The new view, `IntStatView`, can change the underlying data and is to be used in a `RecyclerView`. `RecyclerView`s reuse view holders, like `IntStatView`. If we leave the changed value only in the view itself, it will be lost, or improperly assigned, when the view holder is bound to another data item. For that reason it was necessary to create an auxiliary class, that is mutable, and captures the up-to-date value of the data.